Add files via upload
This assignment's objective is to write a console-based Java application with which a travel agent could present options for travel destinations to a client who wants to redeem his or her accumulated frequent flyer miles.  For the sake of simplicity, we will assume all trips depart from O'Hare International Airport and that the miles represented are required for a roundtrip ticket.  

Depending on the distance, each destination requires a different number of frequent flyer miles to obtain a roundtrip ticket.  Note that, if the client travels during the off-season, he or she may be able to take advantage of the “Fly Cheap” option that requires fewer miles to obtain a roundtrip ticket to a particular destination.