Add files via upload
A parking garage charges a $2.00 minimum fee to park for up to three hours.  The garage charges an additional $0.50 per hour for each hour or part thereof in excess of three hours.  The maximum charge for any given 24-hour period is $10.00.  You may assume that no car parks for longer than 24 hours at a time.

Write an application that calculates and displays the parking charges for each customer who parked in the garage yesterday.  You should prompt the user to enter the hours parked for each customer. 

The program should display the charge for the current customer and should calculate and display the running total of yesterday’s receipts.  After displaying, you should prompt the user asking if they want to enter another customer's parking hours.  It should use the method calculateCharges to determine the charge for each customer.