Add files via upload
As mentioned in the text, a Java identifier consists of one or more characters.  The first character must be an uppercase or lowercase letter of the alphabet, an underscore ( _ ), or a dollar sign ($).  Each remaining character must be an uppercase or lowercase letter of the alphabet, a digit 0-9, an underscore, or a dollar sign.  You  

Write a single Java program named CheckIdentifiers.java that reads in a line of characters, one at a time, and prints out whether the line of characters is a legal Java identifier.  U