Add files via upload
When you and/or your team are finished, submit two – and only two – Java programs, one representing Part 1 and the other Part 2, on Blackboard under In-Class Exercise 1.  Make sure you follow the documentation guidelines and be sure to NOT include the name of a team member that is absent!  The following exercises are taken from the textbook.  All numeric display are to be fully-edited, i.e., floating dollar sign for dollar amounts, commas as necessary and round all numbers to two decimal places.